move exceptions into owning classes
