--- core/src/main/java/ai/doc/tensorio/core/model/ModelException.java
@@ -1,33 +1,0 @@
-/*
- * TIOModelException.java
- * TensorIO
- *
- * Created by Philip Dow on 7/6/2020
- * Copyright (c) 2020 - Present doc.ai (http://doc.ai)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ai.doc.tensorio.core.model;
-
-import androidx.annotation.NonNull;
-
-public class ModelException extends Exception {
-    public ModelException(@NonNull String message) {
-        super(message);
-    }
-
-    public ModelException(@NonNull String message, @NonNull Throwable cause) {
-        super(message, cause);
-    }
-}
